PR Changes - As per the comments
The inbox messages should be parsed when received in the response even if the inbox is not initialized yet. You can initialize it if this is the case. Please check the mobile SDKs implementation.

When messages are updated, it is also checked for any expired messages that need to be deleted.

Call the initialized callback when the user switches.